+++ CHANGELOG.md
@@ -3,7 +3,7 @@ This package was formerly known as `fmutool`.
 
 # Version 1.9.1
 * FIXED: `fmucontainer` handle correctly start/stop in <DefaultExperiment>
-* ADDED: `fmucontainer` able to dynamically change time_step
+* ADDED: `fmucontainer` able to dynamically change time_step. See [container-vr.md](doc/container-vr.md)
 
 # Version 1.9
 * CHANGE: `remoting` code rewrite to (drastically) improve performance